+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/__init__.py
@@ -6,7 +6,7 @@
 
 from ._version import VERSION
 from ._file_client import FileClient
-from ._directory_client import DirectoryClient
+from ._directory_client import ShareDirectoryClient
 from ._share_client import ShareClient
 from ._share_service_client import ShareServiceClient
 from ._shared_access_signature import generate_account_sas, generate_share_sas, generate_file_sas
@@ -38,7 +38,7 @@ __version__ = VERSION
 
 __all__ = [
     'FileClient',
-    'DirectoryClient',
+    'ShareDirectoryClient',
     'ShareClient',
     'ShareServiceClient',
     'ExponentialRetry',