+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/_share_client.py
@@ -30,7 +30,7 @@ from ._generated.models import (
     DeleteSnapshotsOptionType,
     SharePermission)
 from ._deserialize import deserialize_share_properties, deserialize_permission_key, deserialize_permission
-from ._directory_client import DirectoryClient
+from ._directory_client import ShareDirectoryClient
 from ._file_client import FileClient
 
 if TYPE_CHECKING:
@@ -197,21 +197,21 @@ class ShareClient(StorageAccountHostsMixin):
             account_url, share_name=share_name, snapshot=snapshot, credential=credential, **kwargs)
 
     def get_directory_client(self, directory_path=None):
-        # type: (Optional[str]) -> DirectoryClient
+        # type: (Optional[str]) -> ShareDirectoryClient
         """Get a client to interact with the specified directory.
         The directory need not already exist.
 
         :param str directory_path:
             Path to the specified directory.
         :returns: A Directory Client.
-        :rtype: ~azure.storage.file.DirectoryClient
+        :rtype: ~azure.storage.fileshare.ShareDirectoryClient
         """
         _pipeline = Pipeline(
             transport=TransportWrapper(self._pipeline._transport), # pylint: disable = protected-access
             policies=self._pipeline._impl_policies # pylint: disable = protected-access
         )
 
-        return DirectoryClient(
+        return ShareDirectoryClient(
             self.url, share_name=self.share_name, directory_path=directory_path or "", snapshot=self.snapshot,
             credential=self.credential, _hosts=self._hosts, _configuration=self._config, _pipeline=_pipeline,
             _location_mode=self._location_mode)
@@ -652,7 +652,7 @@ class ShareClient(StorageAccountHostsMixin):
 
     @distributed_trace
     def create_directory(self, directory_name, **kwargs):
-        # type: (str, Any) -> DirectoryClient
+        # type: (str, Any) -> ShareDirectoryClient
         """Creates a directory in the share and returns a client to interact
         with the directory.
 
@@ -663,8 +663,8 @@ class ShareClient(StorageAccountHostsMixin):
         :type metadata: dict(str, str)
         :keyword int timeout:
             The timeout parameter is expressed in seconds.
-        :returns: DirectoryClient
-        :rtype: ~azure.storage.file.DirectoryClient
+        :returns: ShareDirectoryClient
+        :rtype: ~azure.storage.fileshare.ShareDirectoryClient
         """
         directory = self.get_directory_client(directory_name)
         kwargs.setdefault('merge_span', True)