+++ sdk/storage/azure-storage-file-share/tests/test_file_client.py
@@ -11,7 +11,7 @@ from azure.storage.fileshare import (
     VERSION,
     ShareServiceClient,
     ShareClient,
-    DirectoryClient,
+    ShareDirectoryClient,
     FileClient)
 
 from filetestcase import (
@@ -25,7 +25,7 @@ from filetestcase import (
 SERVICES = {
     ShareServiceClient: 'file',
     ShareClient: 'file',
-    DirectoryClient: 'file',
+    ShareDirectoryClient: 'file',
     FileClient: 'file',
 }