+++ sdk/storage/azure-storage-file-share/tests/test_file_client_async.py
@@ -12,7 +12,7 @@ from azure.storage.fileshare import VERSION
 from azure.storage.fileshare.aio import (
     ShareServiceClient,
     ShareClient,
-    DirectoryClient,
+    ShareDirectoryClient,
     FileClient)
 
 from filetestcase import (
@@ -26,7 +26,7 @@ from filetestcase import (
 SERVICES = {
     ShareServiceClient: 'file',
     ShareClient: 'file',
-    DirectoryClient: 'file',
+    ShareDirectoryClient: 'file',
     FileClient: 'file',
 }