+++ sdk/storage/azure-storage-file-share/tests/test_share_async.py
@@ -25,7 +25,7 @@ from azure.storage.fileshare import (
 )
 from azure.storage.fileshare.aio import (
     ShareServiceClient,
-    DirectoryClient,
+    ShareDirectoryClient,
     FileClient,
     ShareClient
 )