+++ sdk/storage/azure-storage-file-share/HISTORY.md
@@ -7,7 +7,8 @@
 Going forward, to use this SDK, please install `azure-storage-file-share`.
 Additionally:
 - The namespace within the package has also been renamed to `azure.storage.fileshare`.
-- `FileServiceClient` has been renamed to `ShareServiceClient`
+- `FileServiceClient` has been renamed to `ShareServiceClient`.
+- `DirectoryClient` has been renamed to `ShareDirectoryClient`.
 
 **Breaking changes**